% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRNA.R
\name{intersectRNA}
\alias{intersectRNA}
\title{Intersect two bed files - using GenomicRanges in R}
\usage{
intersectRNA(A, B, stranded = TRUE, A_flank = NA, B_flank = NA)
}
\arguments{
\item{B_flank}{}
}
\description{
Intersect two bed files - using GenomicRanges in R
}
