% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCoverage.R
\name{scale_coverage}
\alias{scale_coverage}
\title{Perform scaling operation on a single vector}
\usage{
scale_coverage(cov, left, right, centre, centre_length)
}
\arguments{
\item{centre_length}{}
}
\description{
Perform scaling operation on a single vector
}
