% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCoverage.R
\name{formatCoverage}
\alias{formatCoverage}
\title{Title}
\usage{
formatCoverage(coverage, smoothing, left = 0, centre = NULL, right = 0,
  centre_length = 1000)
}
\arguments{
\item{centre_length}{}
}
\description{
Title
}
