% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectRNA.R
\name{intersectRNA}
\alias{intersectRNA}
\title{Use pybedtools to intersect two bedfiles and output numerical vectors}
\usage{
intersectRNA(A, B, A_flank = 0, stranded = TRUE)
}
\arguments{
\item{stranded}{}
}
\description{
Use pybedtools to intersect two bedfiles and output numerical vectors
}
